Remove calls of os.Exit from influx cli Run method
One of the first unit tests in the cli tests called the Run method.
Since the Run method called os.Exit, it reported the unit tests as
succeeded. When parallel is set to 1, this skips _all_ unit tests after
the first one. When parallel is set to a higher value, unit tests run by
other processes still get run.

This changes the Run method to return the status as an int. This status
is used to call os.Exit inside of main. That causes the unit tests to
run correctly regardless of how many parallel processes are running.

Also added an additional option to the CLI called `IgnoreSignals`. If
this is set to true, then signals are not registered with the process.
Setting signals doesn't really work in unit tests so it's good to ensure
they don't get set in the first place.

+++ cmd/influx/cli/cli.go
@@ -52,6 +52,7 @@ type CommandLine struct {
 	Path             string
 	Compressed       bool
 	Quit             chan struct{}
+	IgnoreSignals    bool // Ignore signals normally caught by this process (used primarily for testing)
 	osSignals        chan os.Signal
 	historyFile      *os.File
 }
@@ -66,9 +67,11 @@ func New(version string) *CommandLine {
 }
 
 // Run executes the CLI
-func (c *CommandLine) Run() {
+func (c *CommandLine) Run() int {
 	// register OS signals for graceful termination
-	signal.Notify(c.osSignals, os.Kill, os.Interrupt, syscall.SIGTERM, syscall.SIGQUIT, syscall.SIGHUP)
+	if c.IgnoreSignals {
+		signal.Notify(c.osSignals, os.Kill, os.Interrupt, syscall.SIGTERM, syscall.SIGQUIT, syscall.SIGHUP)
+	}
 
 	var promptForPassword bool
 	// determine if they set the password flag but provided no value
@@ -98,14 +101,14 @@ func (c *CommandLine) Run() {
 		fmt.Fprintf(os.Stderr,
 			"Failed to connect to %s\nPlease check your connection settings and ensure 'influxd' is running.\n",
 			c.Client.Addr())
-		return
+		return 0
 	}
 
 	if c.Execute == "" && !c.Import {
 		token, err := c.DatabaseToken()
 		if err != nil {
 			fmt.Fprintf(os.Stderr, "Failed to check token: %s\n", err.Error())
-			return
+			return 1
 		}
 		if token == "" {
 			fmt.Printf(noTokenMsg)
@@ -125,7 +128,7 @@ func (c *CommandLine) Run() {
 		}
 
 		c.Line.Close()
-		os.Exit(0)
+		return 0
 	}
 
 	if c.Import {
@@ -133,7 +136,7 @@ func (c *CommandLine) Run() {
 		u, e := client.ParseConnectionString(path, c.Ssl)
 		if e != nil {
 			fmt.Println(e)
-			return
+			return 1
 		}
 
 		config := v8.NewConfig()
@@ -152,10 +155,10 @@ func (c *CommandLine) Run() {
 		if err := i.Import(); err != nil {
 			fmt.Printf("ERROR: %s\n", err)
 			c.Line.Close()
-			os.Exit(1)
+			return 1
 		}
 		c.Line.Close()
-		os.Exit(0)
+		return 0
 	}
 
 	c.Version()
@@ -178,6 +181,7 @@ func (c *CommandLine) Run() {
 			close(c.Quit)
 		case <-c.Quit:
 			c.exit()
+			return 0
 		default:
 			l, e := c.Line.Prompt("> ")
 			if e == io.EOF {
@@ -824,6 +828,4 @@ func (c *CommandLine) exit() {
 	// release line resources
 	c.Line.Close()
 	c.Line = nil
-	// exit CLI
-	os.Exit(0)
 }
